+++ src/components/general/navigation/AddPaymentButton/index.ts
@@ -1,0 +1,1 @@
+export * from './AddPaymentButton.component';
