+++ src/components/general/navigation/CenterButton/index.ts
@@ -1,0 +1,1 @@
+export * from './CenterButton.component';
