+++ src/components/general/navigation/PescaNavigator/PescaNavigator/index.ts
@@ -1,0 +1,1 @@
+export * from './PescaNavigator.component';
