+++ src/components/general/navigation/PescaNavigator/PescaScreen/index.ts
@@ -1,0 +1,1 @@
+export * from './PescaScreen.component';
