+++ src/components/general/navigation/PescaTabBar/index.ts
@@ -1,0 +1,1 @@
+export * from './PescaTabBar.component';
