+++ src/components/general/navigation/PescaTabIcon/PescaTabIcon.component.tsx
