+++ src/components/general/navigation/PescaTabIcon/index.ts
@@ -1,0 +1,1 @@
+export * from './PescaTabIcon.component';
