+++ src/screens/settings/settings.screen.tsx
@@ -1,9 +1,9 @@
 import { FlyoutHeader } from '@moneyboy/components/general/flyouts/FlyoutHeader';
 import { ListItem } from '@moneyboy/components/general/lists/ListItem';
 import { Separator } from '@moneyboy/components/general/lists/Separator';
-import { createPescaNavigation } from '@moneyboy/components/general/navigation/pesca-navigator/createPescaNavigation';
-import { PescaNavigatorProps } from '@moneyboy/components/general/navigation/pesca-navigator/pescaNavigator';
-import { ScreenComponentProps } from '@moneyboy/components/general/navigation/pesca-navigator/pescaScreen';
+import { createPescaNavigation } from '@moneyboy/components/general/navigation/PescaNavigator/createPescaNavigation';
+import { PescaNavigatorProps } from '@moneyboy/components/general/navigation/PescaNavigator/PescaNavigator';
+import { ScreenComponentProps } from '@moneyboy/components/general/navigation/PescaNavigator/PescaScreen';
 import { Content } from '@moneyboy/components/general/structure/content';
 import { LogoutButton } from '@moneyboy/components/settings/logoutButton';
 import { ThemeSwitch } from '@moneyboy/components/settings/themeSwitch';
